@@ -1,13 +1,16 @@
 package com.hufudb.onedb.core.sql.rel;
 
 import java.util.List;
+import java.util.Stack;
+import java.util.stream.Collectors;
 
 import com.hufudb.onedb.core.data.Header;
 import com.hufudb.onedb.core.sql.expression.OneDBExpression;
-import com.hufudb.onedb.core.sql.expression.OneDBQuery;
+import com.hufudb.onedb.core.sql.schema.OneDBSchema;
+import com.hufudb.onedb.rpc.OneDBCommon.OneDBQueryProto;
 
+import org.apache.calcite.linq4j.tree.Expression;
 import org.apache.calcite.plan.Convention;
-import org.apache.calcite.plan.RelOptTable;
 import org.apache.calcite.rel.RelNode;
 
 public interface OneDBRel extends RelNode {
@@ -16,55 +19,88 @@ public interface OneDBRel extends RelNode {
   Convention CONVENTION = new Convention.Impl("OneDB", OneDBRel.class);
 
   class Implementor {
-    OneDBQuery currentQuery;
+    OneDBSchema schema;
+    OneDBQueryProto.Builder rootContext;
+    OneDBQueryProto.Builder currentContext;
+    Stack<OneDBQueryProto> contextStack;
 
     Implementor() {
-      currentQuery = new OneDBQuery();
+      rootContext = OneDBQueryProto.newBuilder();
+      currentContext = rootContext;
     }
 
-    public OneDBQuery getQuery() {
-      return currentQuery;
+    public OneDBQueryProto getQuery() {
+      return rootContext.build();
     }
 
-    public void visitChild(int ordinal, RelNode input) {
+    public Expression getSchemaExpression() {
+      return schema.getExpression();
+    }
+
+    public void visitChild(RelNode input) {
       ((OneDBRel) input).implement(this);
     }
 
-    public void setTable(RelOptTable table) {
-      this.currentQuery.table = table;
+    public void visitChild(RelNode left, RelNode right) {
+      OneDBQueryProto.Builder parentBuilder = currentContext;
+      OneDBQueryProto.Builder leftBuilder = OneDBQueryProto.newBuilder();
+      OneDBQueryProto.Builder rightBuilder = OneDBQueryProto.newBuilder();
+      currentContext = leftBuilder;
+      ((OneDBRel) left).implement(this);
+      currentContext = rightBuilder;
+      ((OneDBRel) right).implement(this);
+      currentContext = parentBuilder;
     }
 
-    public void setOneDBTable(OneDBTable oneDBTable) {
-      this.currentQuery.oneDBTable = oneDBTable;
-      this.currentQuery.header = oneDBTable.getHeader();
+    // public void visitJoin(BiRel join) {
+    //   OneDBQueryProto.Builder parentBuilder = currentContext;
+    //   OneDBQueryProto.Builder leftBuilder = OneDBQueryProto.newBuilder();
+    //   OneDBQueryProto.Builder rightBuilder = OneDBQueryProto.newBuilder();
+    //   currentContext = leftBuilder;
+    //   ((OneDBRel) join.getLeft()).implement(this);
+    //   currentContext = rightBuilder;
+    //   ((OneDBRel) join.getRight()).implement(this);
+    //   currentContext = parentBuilder;
+    // }
+
+    public void setSchema(OneDBSchema schema) {
+      if (this.schema == null) {
+        this.schema = schema;
+      }
     }
 
+    public void setTableName(String tableName) {
+      currentContext.setTableName(tableName);
+    }
+
+    // todo: pass proto
     public void setSelectExps(List<OneDBExpression> exps) {
-      this.currentQuery.selectExps = exps;
+      currentContext.addAllSelectExp(exps.stream().map(exp -> exp.toProto()).collect(Collectors.toList()));
     }
 
     public void addFilterExps(OneDBExpression exp) {
-      this.currentQuery.filterExps.add(exp);
+      currentContext.addWhereExp(exp.toProto());
     }
 
     public void setAggExps(List<OneDBExpression> exps) {
-      this.currentQuery.aggExps = exps;
+      currentContext.addAllAggExp(exps.stream().map(exp -> exp.toProto()).collect(Collectors.toList()));
     }
 
     public void setOffset(int offset) {
-      this.currentQuery.offset = offset;
+      currentContext.setOffset(offset);
+      // currentBuilder.offset = offset;
     }
 
     public void setFetch(int fetch) {
-      this.currentQuery.fetch = fetch;
+      currentContext.setFetch(fetch);
     }
 
     public Header getHeader() {
-      return this.currentQuery.header;
+      return Header.EMPTY;
     }
 
     public List<OneDBExpression> getCurrentOutput() {
-      return this.currentQuery.selectExps;
+      return currentContext.getSelectExpList().stream().map(exp -> OneDBExpression.fromProto(exp)).collect(Collectors.toList());
     }
   }
 }
