@@ -2,18 +2,25 @@ package com.hufudb.onedb.core.sql.rel;
 
 import java.util.List;
 
+import javax.annotation.Nullable;
+
 import com.google.common.collect.ImmutableList;
 import com.hufudb.onedb.core.data.Header;
 import com.hufudb.onedb.core.sql.expression.OneDBReference;
 import com.hufudb.onedb.core.sql.rule.OneDBRules;
 
 import org.apache.calcite.plan.RelOptCluster;
+import org.apache.calcite.plan.RelOptCost;
 import org.apache.calcite.plan.RelOptPlanner;
 import org.apache.calcite.plan.RelOptRule;
 import org.apache.calcite.plan.RelOptTable;
 import org.apache.calcite.plan.RelTraitSet;
 import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.core.TableScan;
+import org.apache.calcite.rel.metadata.RelMetadataQuery;
+import org.apache.calcite.rel.rules.JoinCommuteRule;
+import org.apache.calcite.rel.rules.JoinPushThroughJoinRule;
+import org.apache.calcite.rel.rules.SortJoinTransposeRule;
 import org.apache.calcite.rel.type.RelDataType;
 
 public class OneDBTableScan extends TableScan implements OneDBRel {
@@ -31,6 +38,11 @@ public class OneDBTableScan extends TableScan implements OneDBRel {
   }
 
   @Override
+  public RelOptCost computeSelfCost(RelOptPlanner planner, RelMetadataQuery mq) {
+    return super.computeSelfCost(planner, mq).multiplyBy(.05);
+  }
+
+  @Override
   public RelNode copy(RelTraitSet traitSet, List<RelNode> inputs) {
     return this;
   }
@@ -44,12 +56,15 @@ public class OneDBTableScan extends TableScan implements OneDBRel {
     for (RelOptRule rule : OneDBRules.RULES) {
       planner.addRule(rule);
     }
+    planner.removeRule(JoinCommuteRule.Config.DEFAULT.toRule());
+    planner.removeRule(JoinPushThroughJoinRule.LEFT);
+    planner.removeRule(JoinPushThroughJoinRule.RIGHT);
   }
 
   @Override
   public void implement(Implementor implementor) {
-    implementor.setOneDBTable(oneDBTable);
-    implementor.setTable(table);
+    implementor.setTableName(oneDBTable.getTableName());
+    implementor.setSchema(oneDBTable.getSchema());
     implementor.setSelectExps(OneDBReference.fromHeader(oneDBTable.getHeader()));
   }
 }
