@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.apache.calcite.adapter.enumerable.EnumerableRel;
 import org.apache.calcite.adapter.enumerable.EnumerableRelImplementor;
-import org.apache.calcite.adapter.enumerable.JavaRowFormat;
 import org.apache.calcite.adapter.enumerable.PhysType;
 import org.apache.calcite.adapter.enumerable.PhysTypeImpl;
 import org.apache.calcite.linq4j.tree.BlockBuilder;
@@ -18,9 +17,6 @@ import org.apache.calcite.plan.RelTraitSet;
 import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.convert.ConverterImpl;
 import org.apache.calcite.rel.metadata.RelMetadataQuery;
-import org.apache.calcite.rel.type.RelDataType;
-
-import com.hufudb.onedb.core.sql.expression.OneDBQuery;
 
 public class OneDBToEnumerableConverter extends ConverterImpl implements EnumerableRel {
 
@@ -41,18 +37,20 @@ public class OneDBToEnumerableConverter extends ConverterImpl implements Enumera
   @Override
   public Result implement(EnumerableRelImplementor implementor, Prefer pref) {
     final OneDBRel.Implementor oImplementor = new OneDBRel.Implementor();
-    oImplementor.visitChild(0, getInput());
+    oImplementor.visitChild(getInput());
     return implement(implementor, oImplementor, pref);
   }
 
   public Result implement(EnumerableRelImplementor implementor, OneDBRel.Implementor oImplementor, Prefer pref) {
     final BlockBuilder builder = new BlockBuilder();
-    final RelDataType rowType = getRowType();
-    final PhysType physType = PhysTypeImpl.of(implementor.getTypeFactory(), rowType, pref.prefer(JavaRowFormat.ARRAY));
-    OneDBQuery oQuery = oImplementor.getQuery();
-    final Expression table = builder.append("table", oQuery.table.getExpression(OneDBTable.OneDBQueryable.class));
-    final Expression query = builder.append("query", Expressions.constant(oQuery.toProto().toString(), String.class));
-    Expression enumerable = builder.append("enumerable", Expressions.call(table, OneDBMethod.ONEDB_TABLE_QUERY.method, query));
+    final PhysType physType = PhysTypeImpl.of(implementor.getTypeFactory(), getRowType(), pref.preferArray());
+    String queryProtoStr = oImplementor.getQuery().toString();
+    // get OneDBSchema for query
+    Expression schema = oImplementor.getSchemaExpression();
+    // get queryContext
+    Expression queryContext = builder.append("queryContext", Expressions.constant(queryProtoStr, String.class));
+    // call the query method on onedbschema
+    Expression enumerable = builder.append("enumerable", Expressions.call(schema, "query", queryContext));
     builder.add(Expressions.return_(null, enumerable));
     return implementor.result(physType, builder.toBlock());
   }
